Rename the span_to_dfw_record module to span_to_dfw. Since this module had a function of the same name and because the function was imported into the package, this made the rest of the module difficult to access
